chore: separate Session Inspector consts from common consts
