+++ app/renderer/components/Inspector/Inspector.js
@@ -17,12 +17,12 @@ import {debounce} from 'lodash';
 import React, {useEffect, useRef, useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 
+import {WINDOW_DIMENSIONS} from '../../constants/COMMON';
 import {
   INSPECTOR_TABS,
   MJPEG_STREAM_CHECK_INTERVAL,
   SESSION_EXPIRY_PROMPT_TIMEOUT,
-  WINDOW_DIMENSIONS,
-} from '../../constants/COMMON';
+} from '../../constants/SESSION_INSPECTOR';
 import {SCREENSHOT_INTERACTION_MODE} from '../../constants/SCREENSHOT';
 import {clipboard} from '../../polyfills';
 import {BUTTON} from '../AntdTypes';
